refactor: isolate gated shared add
